Functionality Implementation

+++ user-service/src/main/java/com/flexe/userservice/service/UserService.java
@@ -1,15 +1,10 @@
 package com.flexe.userservice.service;
-import com.flexe.userservice.entity.user.User;
-import com.flexe.userservice.entity.user.UserAccount;
-import com.flexe.userservice.entity.user.UserDisplay;
-import com.flexe.userservice.entity.user.UserProfile;
+import com.flexe.userservice.entity.user.*;
 import com.flexe.userservice.repository.UserProfileRepository;
 import com.flexe.userservice.repository.UserRepository;
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
-import java.util.Optional;
 
 @Service
 public class UserService {
@@ -39,12 +34,30 @@ public class UserService {
         return userProfileRepository.findByUserId(userId);
     }
 
+    public UserProfile findUserProfileOrThrow(String userId){
+        UserProfile profile = findProfile(userId);
+        if(profile == null) throw new IllegalArgumentException("User Profile Not Found");
+        return profile;
+    }
+
     public UserDisplay findUserDisplay(String userId) {
         User user = userRepository.findById(userId).orElse(null);
+        if(user == null) return null;
         UserProfile profile = userProfileRepository.findByUserId(userId);
-        if(user == null || profile == null) return null;
         return new UserDisplay(user, profile);
     }
+
+    public UserDisplay findUserDisplayByUsername(String username){
+        User user = userRepository.findByUsername(username).orElse(null);
+        if(user == null) return null;
+
+        return new UserDisplay(user, findUserProfileOrThrow(user.getId()));
+    }
+
+    public User saveUser(User user){
+        return userRepository.save(user);
+    }
+
 
     //User Creation
     public UserDisplay onboardUser(UserDisplay user){
@@ -63,13 +76,43 @@ public class UserService {
     }
 
     //User Deletion
-    public UserAccount deleteUserAccount(UserAccount account){
+    public void deleteUserAccount(String userId){
+        UserDisplay account = findUserDisplay(userId);
+        if(account == null){
+            throw new IllegalArgumentException("User Not Found");
+        }
+
         userRepository.delete(account.getUser());
         userProfileRepository.delete(account.getProfile());
         //Send Request To Post Service
 
         //Send Message To Interaction Service
-        userInteractionService.DeleteUserNode(new UserDisplay(account));
-        return account;
+        userInteractionService.DeleteUserNode(account);
     }
+
+    public void FollowUser(String userId, String targetId){
+        UserProfile userProfile = findUserProfileOrThrow(userId);
+        UserProfile targetProfile = findUserProfileOrThrow(targetId);
+
+        userProfile.AddFollowing();
+        targetProfile.AddFollower();
+        userInteractionService.FollowUser(userProfile.getUserId(), targetProfile.getUserId());
+        SaveUserInteractionChanges(userProfile, targetProfile);
+    }
+
+    public void UnfollowUser(String userId, String targetId){
+        UserProfile userProfile = findUserProfileOrThrow(userId);
+        UserProfile targetProfile = findUserProfileOrThrow(targetId);
+
+        userProfile.RemoveFollowing();
+        targetProfile.RemoveFollower();
+        userInteractionService.UnfollowUser(userProfile.getUserId(), targetProfile.getUserId());
+        SaveUserInteractionChanges(userProfile, targetProfile);
+    }
+
+    public void SaveUserInteractionChanges(UserProfile user, UserProfile target){
+        userProfileRepository.save(user);
+        userProfileRepository.save(target);
+    }
+
 }
